refactor(Branch): Extract logic to new BranchService
WIP. functionality working, need to fix tests
